[US21][T2] Delete Account Finnish

+++ flaskProject/resources/client.py
@@ -1,5 +1,7 @@
 from flask_restful import Resource, reqparse
 from models.client_model import ClientModel
+from models.reserved_running_model import ReservedRunningModel
+from models.moto_model import MotoModel
 
 
 class Client(Resource):
@@ -36,13 +38,37 @@ class Client(Resource):
 
         return entry.json(), 201
 
-    def delete(self, client_id):
-        c = ClientModel.query.filter_by(client_id=client_id).first()
-        if c:
-            c.delete_from_db()
-            return {'message': 'Client with ID [{}] correctly deleted.'.format(client_id)}, 200
-        else:
-            return {'message': 'There is no client with ID [{}] .'.format(client_id)}, 404
+    def delete(self, email_client):
+
+        parser = reqparse.RequestParser()
+        parser.add_argument('password', type=str, required=True, help="Password cannot be left blanck")
+        data = parser.parse_args()
+
+        password = data['password']
+        c = ClientModel.find_by_email(email_client)
+        if c: # Si existe el cliente
+            if c.password == password: # Si la password coincide con la del cliente
+                rr = ReservedRunningModel.find_by_client(c.client_id)
+                if(rr is not None): # Si existe una reserva o start afiliada al cliente, indagamos
+                    moto = MotoModel.find_by_id(rr.moto.id)
+                    if(moto.state == "RESERVED"): # Si la moto esta reservada
+                        # Borramos la fila de rr
+                        rr.delete_from_db()
+                        # Actualizamos el estado de la moto
+                        moto.state = "AVAILABLE"
+                        moto.save_to_db()
+                        # Borramos el cliente
+                        c.delete_from_db()
+                        return {"message": "Client DELETED successfully"}, 200
+                    if(moto.state == "ACTIVE"): # No podemos borrar el cliente porque antes tiene que finalizar el trayecto
+                        return {'message': 'The account cannot be deleted because you have a journey started'}, 401
+                else: # Si no existe reserva o start, podemos borrarlo directamente
+                    c.delete_from_db()
+                    return {"message": "Client DELETED successfully"}, 200
+            else: # Password erronea
+                return {'message': 'ERROR: The password is not correct.'}, 400
+        else: # Si no existe el cliente
+            return {'message': 'ERROR: There is no client with email [{}] .'.format(email_client)}, 404
 
     def put(self, client_id):
         # Creamos el request parser, que nos ayudará a manejar la informacion de entrada
